ma_health_forecast/requirements.txt
pandas>=2.0
numpy>=1.26
matplotlib>=3.0
pyyaml>=6.0
statsmodels>=0.14
openpyxl>=3.1
requests>=2.0
google-genai>=0.3.0
python-dotenv>=1.0.0
flask>=3.0
gunicorn>=21.2.0

yfinance>=0.2.0
